deep-thinking/requirements.txt
chess==1.6.1
# comet-ml==3.13.2
# easy-to-hard-data==1.0.0
einops==0.7.0
hydra-core==1.3.2
icecream==2.1.3
matplotlib #~=3.5.0
numpy~=1.26.0
omegaconf==2.3.0
pandas==2.1.3
Pillow
seaborn==0.13.0
tensorboard
tqdm
wandb
torch
torchvision
gymnasium
minigrid
gym
IPython